+++ src/main/java/org/controller/StatsController.java
@@ -3,13 +3,11 @@ package org.controller;
 import atlantafx.base.theme.Styles;
 import javafx.collections.ListChangeListener;
 import javafx.fxml.FXML;
-import javafx.fxml.FXMLLoader;
 import javafx.fxml.Initializable;
 import javafx.scene.control.Button;
 import javafx.scene.control.DatePicker;
 import javafx.scene.control.Label;
 import javafx.scene.layout.VBox;
-import org.app.App;
 import org.context.ControllerRegistry;
 import org.context.GlobalContext;
 import org.model.transaction.Transaction;
@@ -104,6 +102,12 @@ public class StatsController extends VBox implements Initializable {
                 }
                 return transaction.getDate().isAfter(newValue.minusDays(1)) && transaction.getDate().isBefore(toDate.getValue().plusDays(1));
             });
+            GlobalContext.getFilteredDailyPrep().setPredicate(dailyPrepDate -> {
+                if (dailyPrepDate == null) {
+                    return true;
+                }
+                return dailyPrepDate.getDate().isAfter(newValue.minusDays(1)) && dailyPrepDate.getDate().isBefore(toDate.getValue().plusDays(1));
+            });
         });
         toDate.valueProperty().addListener((observable, oldValue, newValue) -> {
             GlobalContext.getFilteredTransactions().setPredicate(transaction -> {
@@ -111,6 +115,12 @@ public class StatsController extends VBox implements Initializable {
                     return true;
                 }
                 return transaction.getDate().isBefore(newValue.plusDays(1)) && transaction.getDate().isAfter(fromDate.getValue().minusDays(1));
+            });
+            GlobalContext.getFilteredDailyPrep().setPredicate(dailyPrepDate -> {
+                if (dailyPrepDate == null) {
+                    return true;
+                }
+                return dailyPrepDate.getDate().isBefore(newValue.plusDays(1)) && dailyPrepDate.getDate().isAfter(fromDate.getValue().minusDays(1));
             });
         });