+++ src/main/java/org/controller/TradeController.java
@@ -2,7 +2,9 @@ package org.controller;
 
 import atlantafx.base.theme.Styles;
 import javafx.fxml.FXML;
+import javafx.fxml.FXMLLoader;
 import javafx.fxml.Initializable;
+import javafx.scene.Node;
 import javafx.scene.control.*;
 import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.scene.layout.VBox;
@@ -45,14 +47,15 @@ public class TradeController extends VBox implements Initializable {
     public TableColumn<Transaction, String> tradeFormation;
     @FXML
     public TableColumn<Transaction, Void> edit;
+    @FXML
+    public Button toolbarDeleteBtn;
 
+    Alert confirmDelete = new Alert(Alert.AlertType.INFORMATION);
     StatsController statsController;
-
     DbManager db = new DbManager();
 
     @Override
     public void initialize(URL location, ResourceBundle resources) {
-
         ///Init Controllers
         ControllerRegistry.register(TradeController.class, this);
         statsController = ControllerRegistry.get(StatsController.class);
@@ -71,51 +74,12 @@ public class TradeController extends VBox implements Initializable {
         tradeCommission.setCellValueFactory(new PropertyValueFactory<Transaction, Double>("commission"));
         tradeFormation.setCellValueFactory(new PropertyValueFactory<Transaction, String>("formation"));
 
-        //Format Column data
-        edit.setCellFactory(col -> new TableCell<>() {
-            private final Button btn = new Button("Delete");
-
-            {
-                btn.getStyleClass().addAll(Styles.BUTTON_OUTLINED, Styles.DANGER);
-                btn.setOnAction(e -> {
-                    Transaction transaction = getTableView().getItems().get(getIndex());
-                    try {
-                        db.setBdConnection();
-                        db.deleteTransaction(transaction);
-                        GlobalContext.replaceMasterList(db.getAllTransactions());
-                        db.closeBdConnection();
-                    } catch (IOException | SQLException ex) {
-                        throw new RuntimeException(ex);
-                    }
-                });
-            }
-
-            @Override
-            protected void updateItem(Void item, boolean empty) {
-                super.updateItem(item, empty);
-                if (empty) {
-                    setGraphic(null);
-                } else {
-                    // Show button only if this row is the selected one
-                    TableView.TableViewSelectionModel<Transaction> selectionModel = getTableView().getSelectionModel();
-                    if (selectionModel.getSelectedIndex() == getIndex()) {
-                        setGraphic(btn);
-                    } else {
-                        setGraphic(null);
-                    }
-
-                    // Update when selection changes
-                    selectionModel.selectedIndexProperty().addListener((obs, oldSel, newSel) -> {
-                        if (newSel.intValue() == getIndex()) {
-                            setGraphic(btn);
-                        } else {
-                            setGraphic(null);
-                        }
-                    });
-                }
-            }
+        //Enable delete btn on select
+        tradesTable.getSelectionModel().selectedItemProperty().addListener((observable, oldValue, newValue) -> {
+            toolbarDeleteBtn.setDisable(false);
         });
 
+        //Format Column data
         tradeProfit.setCellFactory(new Callback<TableColumn<Transaction, Double>, TableCell<Transaction, Double>>() {
             @Override
             public TableCell<Transaction, Double> call(TableColumn<Transaction, Double> param) {
@@ -137,13 +101,47 @@ public class TradeController extends VBox implements Initializable {
 
         ///Style Table
         tradesTable.getStyleClass().add(Styles.STRIPED);
+        toolbarDeleteBtn.setDisable(true);
 
     }
 
     @FXML
     private void addTrade() {
+        Node addTransactionDialog;
+        ///Load Dialog
+        try {
+            addTransactionDialog = new FXMLLoader(getClass().getResource("/org/app/fxml/addTransactionDialog.fxml")).load();
+        } catch (IOException exception) {
+            throw new RuntimeException(exception);
+        }
         MainController mainController = ControllerRegistry.get(MainController.class);
-        mainController.showModal();
+        mainController.showModal(addTransactionDialog);
+    }
+
+    @FXML
+    private void deleteTrade() throws IOException {
+        Transaction transaction = tradesTable.getSelectionModel().getSelectedItem();
+        confirmDelete.setTitle("Delete Trade?");
+        confirmDelete.setHeaderText(null);
+        confirmDelete.setContentText("Are you sure you want to delete trade: \n" +
+                "Date: " + transaction.getDate().toString() + "\n" +
+                "Symbol: " + transaction.getSymbol() + "\n" +
+                "Direction: " + transaction.getDirection() + "\n" +
+                "Open: " + transaction.getOpen() + "\n" +
+                "Close: " + transaction.getClose() + "\n" +
+                "Formation: " + transaction.getFormation() + "\n");
+
+        if (confirmDelete.showAndWait().get() == ButtonType.OK) {
+            try {
+                db.setBdConnection();
+                db.deleteTransaction(transaction);
+                GlobalContext.replaceMasterList(db.getAllTransactions());
+                db.closeBdConnection();
+            } catch (IOException | SQLException ex) {
+                throw new RuntimeException(ex);
+            }
+        }
+        toolbarDeleteBtn.setDisable(true);
     }
 
 }