+++ src/main/java/org/manager/DbManager.java
@@ -1,5 +1,8 @@
 package org.manager;
 
+import org.model.dailyPrep.DailyPrep;
+import org.model.dailyPrep.DailyPrepDate;
+import org.model.dailyPrep.DailyPrepItems;
 import org.model.transaction.Transaction;
 import org.service.SqliteConnection;
 
@@ -37,16 +40,16 @@ public class DbManager {
         }
     }
 
-    public boolean transactionTableExists() {
-        if (isDbConnected()) {
+    public boolean doesTableExist(String tableName, DbManager db) {
+        if (db.isDbConnected()) {
             boolean exists = false;
             try {
                 Statement stmt = bdConnection.createStatement();
-                String sql = "SELECT name FROM sqlite_master WHERE type='table' AND name='" + "transactions" + "'";
+                String sql = "SELECT name FROM sqlite_master WHERE type='table' AND name='" + tableName + "'";
                 ResultSet rs = stmt.executeQuery(sql);
                 exists = rs.next(); // If rs.next() returns true, a row was found, meaning the table exists.
             } catch (SQLException e) {
-                System.err.println("Error checking table existence: " + e.getMessage());
+                System.err.println("Error checking" + tableName + " existence: " + e.getMessage());
                 return exists;
             } finally {
                 return exists;
@@ -56,8 +59,27 @@ public class DbManager {
         }
     }
 
-    public void createTransactionTable() throws SQLException {
-        if (isDbConnected()) {
+    public boolean foreignKeyEnabled(DbManager db) {
+        if (db.isDbConnected()) {
+            boolean exists = false;
+            try {
+                Statement stmt = bdConnection.createStatement();
+                String sql = "PRAGMA foreign_keys;";
+                ResultSet rs = stmt.executeQuery(sql);
+                exists = rs.getInt(1) == 1;
+            } catch (SQLException e) {
+                System.err.println("Error checking foreign_keys existence: " + e.getMessage());
+                return exists;
+            } finally {
+                return exists;
+            }
+        } else {
+            return false;
+        }
+    }
+
+    public void createTransactionTable(DbManager db) throws SQLException {
+        if (db.isDbConnected()) {
             Statement statement = null;
             try {
                 statement = bdConnection.createStatement();
@@ -78,7 +100,68 @@ public class DbManager {
                 e.printStackTrace();
             } finally {
                 statement.close();
-                bdConnection.close();
+            }
+        }
+    }
+
+    public void enableForeignKeys(DbManager db) throws SQLException {
+        if (db.isDbConnected()) {
+            Statement statement = null;
+            try {
+                statement = bdConnection.createStatement();
+                String createTableSQL = "PRAGMA foreign_keys = ON;";
+                statement.execute(createTableSQL);
+            } catch (SQLException e) {
+                e.printStackTrace();
+            } finally {
+                statement.close();
+            }
+        }
+    }
+
+    public void createDailyPrepDateTable(DbManager db) throws SQLException {
+        if (db.isDbConnected()) {
+            Statement statement = null;
+            try {
+                statement = bdConnection.createStatement();
+                String createTableSQL = "CREATE TABLE DailyPrepDate (\n" +
+                        "dailyPrepDateId INTEGER PRIMARY KEY AUTOINCREMENT,\n" +
+                        "date ANY NOT NULL\n" +
+                        ");";
+                statement.execute(createTableSQL);
+            } catch (SQLException e) {
+                e.printStackTrace();
+            } finally {
+                statement.close();
+            }
+        }
+    }
+
+    public void createDailyPrepTable(DbManager db) throws SQLException {
+        if (db.isDbConnected()) {
+            Statement statement = null;
+            try {
+                statement = bdConnection.createStatement();
+                String createTableSQL = "CREATE TABLE DailyPrep (\n" +
+                        "dailyPrepId INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT,\n" +
+                        "date ANY,\n" +
+                        "dailyPrepDateId INTEGER,\n" +
+                        "symbol TEXT,\n" +
+                        "dailyEvents TEXT,\n" +
+                        "hourlyTrend TEXT,\n" +
+                        "halfHourlyTrend TEXT,\n" +
+                        "dailyTrend TEXT,\n" +
+                        "hh_ll_3_bars_high REAL,\n" +
+                        "hh_ll_3_bars_low REAL,\n" +
+                        "hh_ll_any_high REAL,\n" +
+                        "hh_ll_any_low REAL,\n" +
+                        "FOREIGN KEY (dailyPrepDateId) REFERENCES parent_table(DailyPrepDate) ON DELETE CASCADE ON UPDATE NO ACTION\n" +
+                        ");";
+                statement.execute(createTableSQL);
+            } catch (SQLException e) {
+                e.printStackTrace();
+            } finally {
+                statement.close();
             }
         }
     }
@@ -111,7 +194,6 @@ public class DbManager {
             } finally {
                 ps.close();
                 rs.close();
-                bdConnection.close();
             }
         }
         System.out.println("Get All Transactions failed! DB is not connected");
@@ -141,6 +223,28 @@ public class DbManager {
             ps.setDouble(7, close);
             ps.setDouble(8, profit);
             ps.setString(9, formation);
+            ps.executeUpdate();
+            ps.close();
+        } catch (Exception e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    public void addDailyPrepItem(
+            DailyPrepItems dailyPrepItem) throws SQLException {
+        PreparedStatement ps = null;
+        String query = "update DailyPrep set dailyEvents = ?, hourlyTrend = ?, halfHourlyTrend = ?, dailyTrend = ?, hh_ll_3_bars_high = ?, hh_ll_3_bars_low = ?, hh_ll_any_high = ?, hh_ll_any_low = ? WHERE dailyPrepId = ?";
+        try {
+            ps = bdConnection.prepareStatement(query);
+            ps.setString(1, dailyPrepItem.getDailyEvents());
+            ps.setString(2, dailyPrepItem.getHourlyTrend());
+            ps.setString(3, dailyPrepItem.getHalfHourlyTrend());
+            ps.setString(4, dailyPrepItem.getDailyTrend());
+            ps.setDouble(5, dailyPrepItem.getHh_ll_3_bars_high());
+            ps.setDouble(6, dailyPrepItem.getHh_ll_3_bars_low());
+            ps.setDouble(7, dailyPrepItem.getHh_ll_any_high());
+            ps.setDouble(8, dailyPrepItem.getHh_ll_any_low());
+            ps.setInt(9, dailyPrepItem.getDailyPrepId());
             ps.executeUpdate();
             ps.close();
         } catch (Exception e) {
@@ -181,6 +285,167 @@ public class DbManager {
         return null;
     }
 
+    public List<DailyPrep> getAllDailyPrepData() throws SQLException {
+        if (isDbConnected()) {
+            PreparedStatement ps = null;
+            ResultSet rs = null;
+            String query = "select * from DailyPrepDate ORDER BY date ASC";
+            try {
+                ps = bdConnection.prepareStatement(query);
+                rs = ps.executeQuery();
+                List<DailyPrep> dailyPreps = new ArrayList<>();
+                List<DailyPrepDate> dailyPrepDates = new ArrayList<>();
+                while (rs.next()) {
+                    dailyPrepDates.add(new DailyPrepDate(
+                            rs.getInt("dailyPrepDateId"),
+                            rs.getDate("date").toLocalDate()
+                    ));
+                }
+
+                for (DailyPrepDate dailyPrepDate : dailyPrepDates) {
+                    List<DailyPrepItems> dailyPrepItems = getDailyPrepItems(dailyPrepDate.getDailyPrepDateId());
+                    dailyPreps.add(new DailyPrep(dailyPrepDate.getDailyPrepDateId(), dailyPrepDate.getDate(), null, dailyPrepItems));
+                }
+                return dailyPreps;
+            } catch (Exception e) {
+                throw new RuntimeException(e);
+            } finally {
+                ps.close();
+                rs.close();
+            }
+        }
+        System.out.println("Get All Transactions failed! DB is not connected");
+        return null;
+    }
+
+    public List<DailyPrepItems> getDailyPrepItems(int id) throws SQLException {
+        if (isDbConnected()) {
+            PreparedStatement ps = null;
+            ResultSet rs = null;
+            String query = "SELECT *" +
+                    "FROM DailyPrep \n" +
+                    "WHERE dailyPrepDateId = ?;";
+            try {
+                ps = bdConnection.prepareStatement(query);
+                ps.setInt(1, id);
+                rs = ps.executeQuery();
+                List<DailyPrepItems> dailyPrepItems = new ArrayList<>();
+                while (rs.next()) {
+                    dailyPrepItems.add(new DailyPrepItems(
+                            rs.getDate("date").toLocalDate(),
+                            rs.getInt("dailyPrepId"),
+                            rs.getInt("dailyPrepDateId"),
+                            rs.getString("dailyEvents"),
+                            rs.getString("symbol"),
+                            rs.getString("hourlyTrend"),
+                            rs.getString("halfHourlyTrend"),
+                            rs.getString("dailyTrend"),
+                            rs.getDouble("hh_ll_3_bars_high"),
+                            rs.getDouble("hh_ll_3_bars_low"),
+                            rs.getDouble("hh_ll_any_high"),
+                            rs.getDouble("hh_ll_any_low")));
+                }
+                return dailyPrepItems;
+            } catch (Exception e) {
+                throw new RuntimeException(e);
+            }
+        }
+        return null;
+    }
+
+    public DailyPrepItems getDailyPrepItem(int id, String symbol) throws SQLException {
+        if (isDbConnected()) {
+            PreparedStatement ps = null;
+            ResultSet rs = null;
+            String query = "SELECT *" +
+                    "FROM DailyPrep \n" +
+                    "WHERE dailyPrepDateId = ? AND symbol = ?;";
+            try {
+                ps = bdConnection.prepareStatement(query);
+                ps.setInt(1, id);
+                ps.setString(2, symbol);
+                rs = ps.executeQuery();
+                DailyPrepItems dailyPrepItem = null;
+                while (rs.next()) {
+                    dailyPrepItem = new DailyPrepItems(
+                            rs.getDate("date").toLocalDate(),
+                            rs.getInt("dailyPrepId"),
+                            rs.getInt("dailyPrepDateId"),
+                            rs.getString("dailyEvents"),
+                            rs.getString("symbol"),
+                            rs.getString("hourlyTrend"),
+                            rs.getString("halfHourlyTrend"),
+                            rs.getString("dailyTrend"),
+                            rs.getDouble("hh_ll_3_bars_high"),
+                            rs.getDouble("hh_ll_3_bars_low"),
+                            rs.getDouble("hh_ll_any_high"),
+                            rs.getDouble("hh_ll_any_low"));
+                }
+                return dailyPrepItem;
+            } catch (Exception e) {
+                throw new RuntimeException(e);
+            }
+        }
+        return null;
+    }
+
+    public DailyPrepItems addDailyPrepItem(int id, String symbol, LocalDate date) throws SQLException {
+        if (isDbConnected()) {
+            PreparedStatement ps = null;
+            ResultSet rs = null;
+            String query = "insert into DailyPrep(dailyPrepDateId, symbol, date) VALUES(?,?, ?)";
+            try {
+                ps = bdConnection.prepareStatement(query);
+                ps.setInt(1, id);
+                ps.setString(2, symbol);
+                ps.setDate(3, Date.valueOf(date));
+                ps.executeUpdate();
+                ps.close();
+                return getDailyPrepItem(id, symbol);
+            } catch (Exception e) {
+                throw new RuntimeException(e);
+            }
+        }
+        return null;
+    }
+
+    public DailyPrep addDailyPrepDate(LocalDate date) throws SQLException {
+        if (isDbConnected()) {
+            PreparedStatement ps = null;
+            String query = "insert into DailyPrepDate(date) VALUES(?)";
+            try {
+                ps = bdConnection.prepareStatement(query);
+                ps.setDate(1, Date.valueOf(date));
+                ps.executeUpdate();
+                ps.close();
+                DailyPrepDate dailyPrepDate = getDailyPrepDate(date);
+                return new DailyPrep(dailyPrepDate.getDailyPrepDateId(), dailyPrepDate.getDate(), null, new ArrayList<>());
+            } catch (Exception e) {
+                throw new RuntimeException(e);
+            }
+        }
+        return null;
+    }
+
+    public DailyPrepDate getDailyPrepDate(LocalDate date) throws SQLException {
+        PreparedStatement ps = null;
+        ResultSet rs = null;
+        String query = "SELECT * FROM DailyPrepDate WHERE date = ? LIMIT 1;";
+        try {
+            ps = bdConnection.prepareStatement(query);
+            ps.setDate(1, Date.valueOf(date));
+            rs = ps.executeQuery();
+            DailyPrepDate dailyPrepDate = null;
+            while (rs.next()) {
+                dailyPrepDate = new DailyPrepDate(rs.getInt("dailyPrepDateId"), rs.getDate("date").toLocalDate());
+            }
+            ps.close();
+            return dailyPrepDate;
+        } catch (Exception e) {
+            throw new RuntimeException(e);
+        }
+    }
+
     public void deleteTransaction(Transaction transaction) throws SQLException {
         if (isDbConnected()) {
             String sql = "DELETE FROM transactions WHERE id = ?";
@@ -193,6 +458,74 @@ public class DbManager {
             } catch (SQLException e) {
                 e.printStackTrace();
             }
+        }
+    }
+
+    public void deleteDay(int id) throws SQLException {
+        if (isDbConnected()) {
+            String sql = "DELETE FROM DailyPrepDate WHERE dailyPrepDateId = ?";
+            PreparedStatement preparedStatement = null;
+            try {
+                preparedStatement = bdConnection.prepareStatement(sql);
+                preparedStatement.setInt(1, id); // Set the ID value
+                preparedStatement.executeUpdate();
+                preparedStatement.close();
+            } catch (SQLException e) {
+                e.printStackTrace();
+            }
+        }
+    }
+
+    public void deleteSymbolByDay(int id) throws SQLException {
+        if (isDbConnected()) {
+            String sql = "DELETE FROM DailyPrep WHERE dailyPrepDateId = ?";
+            PreparedStatement preparedStatement = null;
+            try {
+                preparedStatement = bdConnection.prepareStatement(sql);
+                preparedStatement.setInt(1, id); // Set the ID value
+                preparedStatement.executeUpdate();
+                preparedStatement.close();
+            } catch (SQLException e) {
+                e.printStackTrace();
+            }
+        }
+    }
+
+    public void deleteSymbol(int id) throws SQLException {
+        if (isDbConnected()) {
+            String sql = "DELETE FROM DailyPrep WHERE dailyPrepId = ?";
+            PreparedStatement preparedStatement = null;
+            try {
+                preparedStatement = bdConnection.prepareStatement(sql);
+                preparedStatement.setInt(1, id); // Set the ID value
+                preparedStatement.executeUpdate();
+                preparedStatement.close();
+            } catch (SQLException e) {
+                e.printStackTrace();
+            }
+        }
+    }
+
+    public void dbStartUpChecks(DbManager db) throws SQLException {
+        boolean transactionTable = doesTableExist("transactions", db);
+        boolean isForeignKeyEnabled = foreignKeyEnabled(db);
+        boolean dailyPrepDate = doesTableExist("DailyPrepDate", db);
+        boolean dailyPrep = doesTableExist("DailyPrep", db);
+
+        if (!transactionTable) {
+            createTransactionTable(db);
+        }
+
+        if (!isForeignKeyEnabled) {
+            enableForeignKeys(db);
+        }
+
+        if (!dailyPrepDate) {
+            createDailyPrepDateTable(db);
+        }
+
+        if (!dailyPrep) {
+            createDailyPrepTable(db);
         }
     }
 }