+++ src/main/java/org/model/dailyPrep/DailyPrepDate.java
@@ -1,0 +1,13 @@
+package org.model.dailyPrep;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+import java.time.LocalDate;
+
+@Getter
+@AllArgsConstructor
+public class DailyPrepDate {
+    int dailyPrepDateId;
+    LocalDate date;
+}